requirements.txt
imutils==0.5.4
Pillow==9.5.0
opencv-python==4.6.0


